+++ src/Newbe.Mahua.Amanda/Commands/GetQQWalletDataCommand.cs
@@ -1,4 +1,5 @@
 ﻿using Newbe.Mahua.Commands;
+using System.Runtime.Serialization;
 
 namespace Newbe.Mahua.Amanda.Commands
 {
@@ -10,13 +11,25 @@ namespace Newbe.Mahua.Amanda.Commands
         }
     }
 
-    internal class GetQqWalletDataCommand : AmandaCommand
+    [DataContract]
+    public class GetQqWalletDataCommand : AmandaCommand
     {
+        [DataMember]
         public string Type { get; set; }
+
+        [DataMember]
         public string Fromgroup { get; set; }
+
+        [DataMember]
         public string Fromqq { get; set; }
+
+        [DataMember]
         public string Money { get; set; }
+
+        [DataMember]
         public string FriendRemark { get; set; }
+
+        [DataMember]
         public string OrderNo { get; set; }
     }
 }