+++ src/Newbe.Mahua.CQP/Commands/AddFriendRequestCommand.cs
@@ -2,6 +2,7 @@
 using Newbe.Mahua.MahuaEvents;
 using System;
 using System.Collections.Generic;
+using System.Runtime.Serialization;
 
 namespace Newbe.Mahua.CQP.Commands
 {
@@ -26,10 +27,16 @@ namespace Newbe.Mahua.CQP.Commands
         }
     }
 
-    internal class AddFriendRequestCommand : CqpCommand
+    [DataContract]
+    public class AddFriendRequestCommand : CqpCommand
     {
+        [DataMember]
         public DateTime SendTime { get; set; }
+
+        [DataMember]
         public long FromQq { get; set; }
+
+        [DataMember]
         public string Message { get; set; }
     }
 }