+++ src/Newbe.Mahua.CQP/Commands/GroupJoiningInvitationCommand.cs
@@ -2,6 +2,7 @@
 using Newbe.Mahua.MahuaEvents;
 using System;
 using System.Collections.Generic;
+using System.Runtime.Serialization;
 
 namespace Newbe.Mahua.CQP.Commands
 {
@@ -30,12 +31,22 @@ namespace Newbe.Mahua.CQP.Commands
         }
     }
 
-    internal class GroupJoiningInvitationCommand : CqpCommand
+    [DataContract]
+    public class GroupJoiningInvitationCommand : CqpCommand
     {
+        [DataMember]
         public string GroupJoiningInvitationId { get; set; }
+
+        [DataMember]
         public DateTime SendTime { get; set; }
+
+        [DataMember]
         public long ToGroup { get; set; }
+
+        [DataMember]
         public long FromQq { get; set; }
+
+        [DataMember]
         public string Message { get; set; }
     }
 }