+++ src/Newbe.Mahua.CQP/Commands/GroupMessageCommand.cs
@@ -2,6 +2,7 @@
 using Newbe.Mahua.MahuaEvents;
 using System;
 using System.Collections.Generic;
+using System.Runtime.Serialization;
 
 namespace Newbe.Mahua.CQP.Commands
 {
@@ -27,12 +28,22 @@ namespace Newbe.Mahua.CQP.Commands
         }
     }
 
-    internal class GroupMessageCommand : CqpCommand
+    [DataContract]
+    public class GroupMessageCommand : CqpCommand
     {
+        [DataMember]
         public DateTime SendTime { get; set; }
+
+        [DataMember]
         public long GroupNum { get; set; }
+
+        [DataMember]
         public long FromQq { get; set; }
+
+        [DataMember]
         public string FromAnonymous { get; set; }
+
+        [DataMember]
         public string Message { get; set; }
     }
 }