+++ src/Newbe.Mahua.PluginLoader/Commands/MahuaCommand.cs
@@ -1,15 +1,15 @@
 ﻿using MediatR;
-using System;
+using System.Runtime.Serialization;
 
 namespace Newbe.Mahua.Commands
 {
-    [Serializable]
-    public abstract class MahuaCommand<TResult> : IRequest<TResult> where TResult : MahuaCommandResult
+    [DataContract]
+    public class MahuaCommand<TResult> : IRequest<TResult> where TResult : MahuaCommandResult
     {
     }
 
-    [Serializable]
-    public abstract class MahuaCommand : IRequest
+    [DataContract]
+    public class MahuaCommand : IRequest
     {
     }
 }