+++ src/build.ps1
@@ -42,7 +42,7 @@ Task Pack -depends Build -Description "打包" {
     }
 }
 Task NugetPushLocal -depends Pack -Description "推送nuget包到本地" {
-    Get-ChildItem $releaseDir *.nupkg | ForEach-Object { 
+    Get-ChildItem $releaseDir *.nupkg | ForEach-Object {
         Exec {
             cmd /c "$nugetexe push $releaseDir$_ Admin:Admin -Source http://localhost:81/nuget/NewbeGet/"
         }
@@ -50,7 +50,7 @@ Task NugetPushLocal -depends Pack -Description "推送nuget包到本地" {
 }
 
 Task NugetPushNuget -depends Pack -Description "推送nuget包到nuget.org" {
-    Get-ChildItem $releaseDir *.nupkg | ForEach-Object { 
+    Get-ChildItem $releaseDir *.nupkg | ForEach-Object {
         Exec {
             cmd /c "$nugetexe push $releaseDir$_ -Source https://www.nuget.org/api/v2/package"
         }