+++ ReleaseNotes.md
@@ -9,6 +9,16 @@
 
 对于修订号和构建号此列表将不会记录，直接升级没有影响。
 
+## 1.3.0
+
+使用MessagePack替代JSON序列化，消息处理速度明显提升。
+
+## 1.2.0
+
+全新的打包机制：[Newbe.Mahua.Tools.Psake](https://www.nuget.org/packages/Newbe.Mahua.Tools.Psake/)
+被废弃的打包机制：Newbe.Mahua.Msbuild
+[新版快速指南](http://www.newbe.cf/docs/mahua/2017/10/26/Begin-First-Plugin-With-Mahua-In-v1.2.html)
+
 ## 1.1
 
 [#12](https://github.com/Newbe36524/Newbe.Mahua.Framework/issues/12) 提供对群列表、群成员列表的结构化接口支持